+++ samples/kotlin_compose/src/main/java/io/customer/android/sample/kotlin_compose/MainApplication.kt
@@ -5,7 +5,6 @@ import dagger.hilt.android.HiltAndroidApp
 import io.customer.android.sample.kotlin_compose.data.models.setValuesFromBuilder
 import io.customer.android.sample.kotlin_compose.data.repositories.PreferenceRepository
 import io.customer.android.sample.kotlin_compose.data.sdk.InAppMessageEventListener
-import io.customer.android.sample.kotlin_compose.util.ActivityLifecycleEventsListener
 import io.customer.messaginginapp.MessagingInAppModuleConfig
 import io.customer.messaginginapp.ModuleMessagingInApp
 import io.customer.messagingpush.MessagingPushModuleConfig
